Update in the code and some data
Up to now here develop a small library with the required functionalities to later easily automate the training,evaluation and results storing processes. Main execution flow is contained in train_test_model.py. Thereby we have only developed up to the part in which we dinamically choose the train, test and validation datasets based on the test days and a stratified random shuffled split for the validation set